Reorganize polaris.viz
This merge:
* makes a shared `use_mplstyle()` function for loading and using
  the `matplotlib` style file
* moves horizontal visualizaiton to its own module to avoid
  circular imports
* adds the 3 public visualization functions to the main `polaris.viz`
  file for easier importing.
